locallibrary/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==4.2.20
django-filter==24.3
djangorestframework==3.15.2
docopt==0.6.2
idna==3.10
importlib-metadata==8.5.0
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
psycopg2==2.9.10
pygments==2.19.1
python-dateutil==2.9.0.post0
pythonanywhere==0.15.5
pythonanywhere-core==0.2.4
requests==2.32.3
rich==14.0.0
schema==0.7.7
shellingham==1.5.4
six==1.17.0
snakesay==0.10.3
sqlparse==0.5.3
tabulate==0.9.0
typer==0.15.2
typing-extensions==4.13.1
urllib3==2.2.3
zipp==3.20.2
